requirements.txt
Flask>=3.0.0
Pillow>=10.2.0
numpy>=1.26.0
cryptography>=41.0.0
